edit funtion rectified

+++ CRUD-Dapper/Repository/IGenericRepository.cs
@@ -3,8 +3,8 @@
     public interface IGenericRepository<T> where T : class
     {
         Task<IEnumerable<T>> GetAll(string tableName);
-        Task<T> GetById(string tableName, Int64 id);
-        Task Delete(string tableName, Int64 id);
+        Task<T> GetById(string tableName, int id);
+        Task Delete(string tableName, int id);
         Task Add(string tableName, T entity);
         Task Update(string tableName, T entity);
     }
